+++ Source/Anvil/Properties/AssemblyInfo.cs
@@ -2,8 +2,8 @@
 using System.Runtime.InteropServices;
 using System.Windows;
 
-[assembly: AssemblyVersion("0.0.1.5")]
-[assembly: AssemblyFileVersion("0.0.1.5")]
+[assembly: AssemblyVersion("0.0.1.6")]
+[assembly: AssemblyFileVersion("0.0.1.6")]
 [assembly: AssemblyTitle("Anvil")]
 [assembly: AssemblyDescription("A program launcher with configurable environment variables.")]
 [assembly: AssemblyConfiguration("")]